+++ Airship/Common/UAObservable.h
@@ -38,6 +38,6 @@ __attribute__((deprecated("As of version 2.1")))
 - (void)addObserver:(id)observer;
 - (void)removeObserver:(id)observer;
 - (void)removeObservers;
-- (int)countObservers;
+- (NSUInteger)countObservers;
 
 @end
